requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai>=1.12.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.0
httpx==0.25.2
aiofiles==23.2.1
